+++ Flow/test/utopia/flow/test/TestContext.scala
@@ -15,6 +15,6 @@ import scala.concurrent.ExecutionContext
 object TestContext
 {
 	implicit val logger: Logger = SysErrLogger
-	implicit val exc: ExecutionContext = new ThreadPool("Test").executionContext
+	implicit val exc: ExecutionContext = new ThreadPool("Test")
 	implicit val weekdays: WeekDays = MondayToSunday
 }