+++ Flow/test/utopia/flow/test/file/FileSearchTest.scala
@@ -2,10 +2,9 @@ package utopia.flow.test.file
 
 import utopia.flow.async.AsyncExtensions._
 import utopia.flow.async.context.ThreadPool
-import utopia.flow.parse.file.search.Guild
-import utopia.flow.generic.model.mutable.DataType
-import utopia.flow.parse.file.FileExtensions._
 import utopia.flow.collection.CollectionExtensions._
+import utopia.flow.parse.file.FileExtensions._
+import utopia.flow.parse.file.search.Guild
 import utopia.flow.util.logging.{Logger, SysErrLogger}
 
 import scala.concurrent.ExecutionContext
@@ -19,7 +18,7 @@ object FileSearchTest extends App
 {
 	
 	implicit val logger: Logger = SysErrLogger
-	implicit val exc: ExecutionContext = new ThreadPool("File Search Test").executionContext
+	implicit val exc: ExecutionContext = new ThreadPool("File Search Test")
 	val resultFuture = Guild.explore(".") { dir =>
 		dir.children.flatMap {
 			_.filter { _.isRegularFile }