+++ Flow/test/utopia/flow/test/time/WaitTest2.scala
@@ -21,7 +21,7 @@ object WaitTest2 extends App
 {
 	
 	implicit val logger: Logger = SysErrLogger
-	implicit val exc: ExecutionContext = new ThreadPool("test").executionContext
+	implicit val exc: ExecutionContext = new ThreadPool("test")
 	
 	println("Running WaitTest2...")